pyproject.toml
[tool.poetry]
name = "timelapse"
version = "0.2.0"
description = ""
authors = ["Kevin Houlihan <kevin@crimsoncookie.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7"
python-xlib = "^0.27"
pyscreenshot = "^2.2"
pillow = "^7.1.2"

[tool.poetry.scripts]
timelapse = "timelapse:_main"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
